# apps/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AdminIngredientResponse" : {
        "properties" : {
          "abv" : {
            "type" : "number"
          },
          "aliases" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "category" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "domesticAvailability" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "isApproved" : {
            "type" : "boolean"
          },
          "nameEn" : {
            "type" : "string"
          },
          "nameKo" : {
            "type" : "string"
          },
          "priceBand" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "substituteNote" : {
            "type" : "string"
          }
        },
        "required" : [ "aliases", "category", "domesticAvailability", "id", "isApproved", "nameEn", "nameKo", "slug" ],
        "type" : "object"
      },
...
      "BaseSpirit" : {
        "description" : "축 1 · 기주 (단일값 필수, R-C-1)",
        "enum" : [ "gin", "vodka", "whisky", "rum", "agave", "brandy", "liqueur", "wine", "korean", "non-alcoholic" ],
        "type" : "string",
        "x-labels" : {
          "agave" : "데킬라 · 메즈칼",
          "brandy" : "브랜디",
          "gin" : "진",
          "korean" : "전통주",
          "liqueur" : "리큐르",
          "non-alcoholic" : "무알콜",
          "rum" : "럼",
          "vodka" : "보드카",
          "whisky" : "위스키",
          "wine" : "와인 · 스파클링"
        }
      },
...
      "BrandItem" : {
        "properties" : {
          "isSponsored" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          },
          "purchaseUrl" : {
            "type" : "string"
          },
          "requiresAdLabel" : {
            "type" : "boolean"
          }
        },
        "required" : [ "isSponsored", "name", "requiresAdLabel" ],
        "type" : "object"
      },
...
      "CocktailListItem" : {
        "properties" : {
          "abv" : {
            "type" : "number"
          },
          "aromaTags" : {
            "items" : {
              "$ref" : "#/components/schemas/FlavorKey"
            },
            "type" : "array"
          },
          "baseSpirit" : {
            "$ref" : "#/components/schemas/BaseSpirit"
          },
          "glassType" : {
            "type" : "string"
          },
          "isClassic" : {
            "type" : "boolean"
          },
          "method" : {
            "$ref" : "#/components/schemas/Technique"
          },
          "nameEn" : {
            "type" : "string"
          },
          "nameKo" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "stylePrimary" : {
            "$ref" : "#/components/schemas/StyleKey"
          },
          "styles" : {
            "items" : {
              "$ref" : "#/components/schemas/StyleKey"
            },
            "type" : "array"
          },
          "summary" : {
            "type" : "string"
          },
          "sweetness" : {
            "$ref" : "#/components/schemas/SweetLevel"
          }
        },
        "required" : [ "aromaTags", "baseSpirit", "glassType", "isClassic", "method", "nameEn", "nameKo", "slug", "stylePrimary", "styles", "summary", "sweetness" ],
        "type" : "object"
      },
...
      "FacetCounts" : {
        "properties" : {
          "abv" : {
            "additionalProperties" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "object"
          },
          "base" : {
            "additionalProperties" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "object"
          },
          "flavor" : {
            "additionalProperties" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "object"
          },
          "method" : {
            "additionalProperties" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "object"
          },
          "style" : {
            "additionalProperties" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "object"
          },
          "sweet" : {
            "additionalProperties" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "object"
          }
        },
        "required" : [ "abv", "base", "flavor", "method", "style", "sweet" ],
        "type" : "object"
      },
      "FlavorKey" : {
        "description" : "향 태그 1~3개 (R-F1.2-1). 카테고리가 아니다",
        "enum" : [ "citrus", "sour", "fruity", "floral", "herbal", "spicy", "smoky", "bitter", "nutty", "creamy" ],
        "type" : "string",
        "x-labels" : {
          "bitter" : "쓴맛",
          "citrus" : "시트러스",
          "creamy" : "크리미",
          "floral" : "플로럴",
          "fruity" : "프루티",
          "herbal" : "허브",
          "nutty" : "너티",
          "smoky" : "스모키",
          "sour" : "산미",
          "spicy" : "스파이시"
        }
      },
...
      "IngredientCocktailItem" : {
        "properties" : {
          "isOptional" : {
            "type" : "boolean"
          },
          "nameEn" : {
            "type" : "string"
          },
          "nameKo" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          }
        },
        "required" : [ "isOptional", "nameEn", "nameKo", "slug", "summary" ],
        "type" : "object"
      },
      "IngredientDetail" : {
        "properties" : {
          "abv" : {
            "type" : "number"
          },
          "aliases" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "brands" : {
            "items" : {
              "$ref" : "#/components/schemas/BrandItem"
            },
            "type" : "array"
          },
          "category" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "domesticAvailability" : {
            "type" : "string"
          },
          "nameEn" : {
            "type" : "string"
          },
          "nameKo" : {
            "type" : "string"
          },
          "priceBand" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "substituteNote" : {
            "type" : "string"
          }
        },
        "required" : [ "aliases", "brands", "category", "domesticAvailability", "nameEn", "nameKo", "slug" ],
        "type" : "object"
      },
      "IngredientItem" : {
        "properties" : {
          "abv" : {
            "type" : "number"
          },
          "category" : {
            "type" : "string"
          },
          "domesticAvailability" : {
            "type" : "string"
          },
          "nameEn" : {
            "type" : "string"
          },
          "nameKo" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          }
        },
        "required" : [ "category", "domesticAvailability", "nameEn", "nameKo", "slug" ],
        "type" : "object"
      },
...
      "PageMeta" : {
        "properties" : {
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "number", "size", "totalElements", "totalPages" ],
        "type" : "object"
      },
      "PageQuery" : {
        "properties" : {
          "offset" : {
            "format" : "int64",
            "type" : "integer"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "items" : {
              "$ref" : "#/components/schemas/SortOrder"
            },
            "type" : "array"
          }
        },
        "required" : [ "offset", "page", "size", "sort" ],
        "type" : "object"
      },
...
      "PageResponseCocktailListItem" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/CocktailListItem"
            },
            "type" : "array"
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMeta"
          }
        },
        "required" : [ "items", "page" ],
        "type" : "object"
      },
      "PageResponseIngredientCocktailItem" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/IngredientCocktailItem"
            },
            "type" : "array"
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMeta"
          }
        },
        "required" : [ "items", "page" ],
        "type" : "object"
      },
      "PageResponseIngredientItem" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/IngredientItem"
            },
            "type" : "array"
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMeta"
          }
        },
        "required" : [ "items", "page" ],
        "type" : "object"
      },
...
      "PublishResponse" : {
        "properties" : {
          "publishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "required" : [ "slug", "status" ],
        "type" : "object"
      },
...
      "SortOrder" : {
        "properties" : {
          "ascending" : {
            "type" : "boolean"
          },
          "property" : {
            "type" : "string"
          }
        },
        "required" : [ "ascending", "property" ],
        "type" : "object"
      },
...
      "StyleKey" : {
        "description" : "축 2 · 스타일 (복수, style_primary 필수)",
        "enum" : [ "highball", "sour", "spirit-forward", "spritz", "tiki", "creamy", "hot", "frozen", "shot" ],
        "type" : "string",
        "x-labels" : {
          "creamy" : "크리미",
          "frozen" : "프로즌",
          "highball" : "하이볼",
          "hot" : "핫",
          "shot" : "샷",
          "sour" : "사워",
          "spirit-forward" : "스피릿 포워드",
          "spritz" : "스프리츠",
          "tiki" : "티키"
        }
      },
...
      "SweetLevel" : {
        "description" : "당도 4단계",
        "enum" : [ "dry", "semi_dry", "semi_sweet", "sweet" ],
        "type" : "string",
        "x-labels" : {
          "dry" : "드라이",
          "semi_dry" : "세미 드라이",
          "semi_sweet" : "세미 스위트",
          "sweet" : "스위트"
        }
      },
...
      "Technique" : {
        "description" : "축 3 · 메이킹 방법 (단일값 필수)",
        "enum" : [ "build", "shake", "stir", "blend", "etc" ],
        "type" : "string",
        "x-labels" : {
          "blend" : "블렌드",
          "build" : "잔에서 조립",
          "etc" : "기타",
          "shake" : "흔들어 섞기",
          "stir" : "휘저어 섞기"
        }
      },
...
      "ValidationProblem" : {
        "description" : "도메인 규칙 위반. `violations` 에 실패한 항목을 **전부** 담는다 (FR-ADMIN-003)",
        "properties" : {
          "detail" : {
            "description" : "사람이 읽을 설명",
            "type" : "string"
          },
          "instance" : {
            "description" : "요청 경로",
            "example" : "/api/v1/admin/cocktails/1/publish",
            "type" : "string"
          },
          "status" : {
            "description" : "HTTP 상태",
            "example" : 422,
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "description" : "짧은 제목",
            "example" : "발행 조건 미충족",
            "type" : "string"
          },
          "type" : {
            "description" : "RFC 7807 문제 유형 URI",
            "example" : "/problems/validation",
            "type" : "string"
          },
          "violations" : {
            "description" : "실패한 항목 전부. 비어 있지 않다",
            "items" : {
              "$ref" : "#/components/schemas/Violation"
            },
            "type" : "array"
          }
        },
        "required" : [ "status", "title", "type", "violations" ],
        "type" : "object"
      },
...
      "Violation" : {
        "description" : "실패한 항목 전부. 비어 있지 않다",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "message" ],
        "type" : "object"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/api/v1/admin/cocktails/{id}/archive" : {
      "post" : {
        "description" : "published → archived.",
        "operationId" : "archive",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublishResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "보관",
        "tags" : [ "admin-cocktail-controller" ]
      }
    },
    "/api/v1/admin/cocktails/{id}/publish" : {
      "post" : {
        "description" : "게이트 실패는 422 + violations 전부. 이미 발행됐으면 409.",
        "operationId" : "publish",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "422" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description" : "발행 게이트 실패. violations 에 남은 조건이 전부 담긴다"
          }
        },
        "summary" : "발행",
        "tags" : [ "admin-cocktail-controller" ]
      }
    },
...
    "/api/v1/admin/cocktails/{id}/unpublish" : {
      "post" : {
        "description" : "published → draft. 게이트를 검사하지 않는다.",
        "operationId" : "unpublish",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublishResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "회수",
        "tags" : [ "admin-cocktail-controller" ]
      }
    },
...
    "/api/v1/admin/ingredients/{id}" : {
      "get" : {
        "operationId" : "find_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminIngredientResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "재료 조회 (미승인 포함)",
        "tags" : [ "admin-ingredient-controller" ]
      }
    },
    "/api/v1/admin/ingredients/{id}/approve" : {
      "post" : {
        "description" : "admin 만 가능하다 (SPEC-08 §2). 재승인은 409. 감사에 남는다.",
        "operationId" : "approve",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminIngredientResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "재료 승인",
        "tags" : [ "admin-ingredient-controller" ]
      }
    },
...
    "/api/v1/cocktails" : {
      "get" : {
        "description" : "발행분만 반환한다. 필터 결과는 색인하지 않는다 (X-Robots-Tag: noindex).",
        "operationId" : "list_2",
        "parameters" : [ {
          "description" : "기주 슬러그. 콤마로 여러 개 — **OR**",
          "in" : "query",
          "name" : "base",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "스타일 슬러그. 콤마로 여러 개 — **OR**. style_primary 가 아니라 보유 스타일 전체와 맞춘다",
          "in" : "query",
          "name" : "style",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "메이킹 방법 슬러그. 콤마로 여러 개 — **OR**",
          "in" : "query",
          "name" : "method",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "당도 슬러그. **단일값** — 여러 개를 주면 400",
          "in" : "query",
          "name" : "sweet",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "도수 구간 na·low·mid·high. 콤마로 여러 개 — **OR** (ADR-0003)",
          "in" : "query",
          "name" : "abv",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "향·맛 슬러그. 콤마로 여러 개 — **AND**. 전부 가진 것만 (SPEC-07 §3.1)",
          "in" : "query",
          "name" : "flavor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "이름 검색. 초성·별칭까지 보는 정밀 검색은 /search 다",
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/PageQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseCocktailListItem"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "칵테일 목록 · 필터",
        "tags" : [ "cocktail-list-controller" ]
      }
    },
    "/api/v1/cocktails/facets" : {
      "get" : {
        "description" : "기주·스타일·메이킹·당도·도수는 같은 축 선택을 무시하고, 향·맛만 현재 선택에 더했을 때의 수다 (SPEC-07 §3.2).",
        "operationId" : "facets",
        "parameters" : [ {
          "description" : "기주 슬러그. 콤마로 여러 개",
          "in" : "query",
          "name" : "base",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "style",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "method",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "sweet",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "abv",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "flavor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FacetCounts"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "필터 값별 결과 개수",
        "tags" : [ "facet-controller" ]
      }
    },
...
    "/api/v1/ingredients" : {
      "get" : {
        "description" : "**승인된 재료만** 반환한다 (FR-INGREDIENT-001).",
        "operationId" : "list_1",
        "parameters" : [ {
          "description" : "재료 카테고리 슬러그 (7종)",
          "in" : "query",
          "name" : "category",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "국내 유통 슬러그 (4종) — PRIN-P05",
          "in" : "query",
          "name" : "availability",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/PageQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseIngredientItem"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "재료 사전 목록",
        "tags" : [ "ingredient-controller" ]
      }
    },
    "/api/v1/ingredients/{slug}" : {
      "get" : {
        "description" : "FR-INGREDIENT-002 의 6개 항목. 브랜드의 isSponsored 는 항상 실려 나간다 (NFR-L-02).",
        "operationId" : "detail",
        "parameters" : [ {
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IngredientDetail"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "재료 상세",
        "tags" : [ "ingredient-controller" ]
      }
    },
    "/api/v1/ingredients/{slug}/cocktails" : {
      "get" : {
        "description" : "표준 레시피 · 발행분만. 대체재로만 등장하는 것은 제외한다 (R-F1.3-1).",
        "operationId" : "cocktails",
        "parameters" : [ {
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/PageQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseIngredientCocktailItem"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "이 재료를 쓰는 칵테일",
        "tags" : [ "ingredient-cocktails-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}